+++ EtLast.ConsoleHost/EtlContextSerilogAdapter.cs
@@ -321,7 +321,7 @@ internal class EtlContextSerilogAdapter : IEtlContextListener
         }
     }
 
-    public void OnContextIoCommandStart(int uid, IoCommandKind kind, string location, string path, IProcess process, int? timeoutSeconds, string command, string transactionId, Func<IEnumerable<KeyValuePair<string, object>>> argumentListGetter, string message, string messageExtra = null)
+    public void OnContextIoCommandStart(long uid, IoCommandKind kind, string location, string path, IProcess process, int? timeoutSeconds, string command, string transactionId, Func<IEnumerable<KeyValuePair<string, object>>> argumentListGetter, string message, string messageExtra = null)
     {
         if (_ioLogger == null)
             return;
@@ -346,7 +346,7 @@ internal class EtlContextSerilogAdapter : IEtlContextListener
         _ioLogger.Write(LogEventLevel.Information, sb.ToString());
     }
 
-    public void OnContextIoCommandEnd(IProcess process, int uid, IoCommandKind kind, long? affectedDataCount, Exception ex)
+    public void OnContextIoCommandEnd(IProcess process, long uid, IoCommandKind kind, long? affectedDataCount, Exception ex)
     {
         if (_ioLogger == null)
             return;
@@ -387,11 +387,11 @@ internal class EtlContextSerilogAdapter : IEtlContextListener
     {
     }
 
-    public void OnSinkStarted(int sinkUid, string location, string path)
+    public void OnSinkStarted(long sinkUid, string location, string path)
     {
     }
 
-    public void OnWriteToSink(IReadOnlyRow row, int sinkUid)
+    public void OnWriteToSink(IReadOnlyRow row, long sinkUid)
     {
     }
