+++ EtLast.Diagnostics.Interface/Events/ProcessInvocationEndEvent.cs
@@ -2,7 +2,7 @@
 
 public class ProcessInvocationEndEvent : AbstractEvent
 {
-    public int InvocationUID { get; set; }
+    public long InvocationUID { get; set; }
     public long ElapsedMilliseconds { get; set; }
     public long? NetTimeMilliseconds { get; set; }
 }
