+++ EtLast.Diagnostics.Interface/Events/ProcessInvocationStartEvent.cs
@@ -2,12 +2,12 @@
 
 public class ProcessInvocationStartEvent : AbstractEvent
 {
-    public int InvocationUID { get; set; }
-    public int InstanceUID { get; set; }
-    public int InvocationCounter { get; set; }
+    public long InvocationUID { get; set; }
+    public long InstanceUID { get; set; }
+    public long InvocationCounter { get; set; }
     public string Type { get; set; }
     public string Kind { get; set; }
     public string Name { get; set; }
     public string Topic { get; set; }
-    public int? CallerInvocationUID { get; set; }
+    public long? CallerInvocationUID { get; set; }
 }
