+++ EtLast.Diagnostics.Interface/Events/RowCreatedEvent.cs
@@ -2,6 +2,6 @@
 
 public class RowCreatedEvent : AbstractRowEvent
 {
-    public int ProcessInvocationUid { get; set; }
+    public long ProcessInvocationUid { get; set; }
     public KeyValuePair<string, object>[] Values { get; set; }
 }
