+++ EtLast.Diagnostics.Interface/Events/SinkStartedEvent.cs
@@ -2,7 +2,7 @@
 
 public class SinkStartedEvent : AbstractRowEvent
 {
-    public int UID { get; set; }
+    public long UID { get; set; }
     public string Location { get; set; }
     public string Path { get; set; }
 }
