+++ EtLast.Diagnostics.Interface/Session/TrackedSink.cs
@@ -3,12 +3,12 @@
 [DebuggerDisplay("{Path}")]
 public class TrackedSink
 {
-    public int UID { get; }
+    public long UID { get; }
     public string Location { get; }
     public string Path { get; }
-    public int RowCount { get; set; }
+    public long RowCount { get; set; }
 
-    public TrackedSink(int uid, string location, string path)
+    public TrackedSink(long uid, string location, string path)
     {
         UID = uid;
         Location = location;
