+++ EtLast.EPPlus/Mutators/EpPlusSingleExcelFileWriterMutator.cs
@@ -11,7 +11,7 @@ public sealed class EpPlusSingleExcelFileWriterMutator<TState> : AbstractMutator
 
     private TState _state;
     private ExcelPackage _package;
-    private int? _sinkUid;
+    private long? _sinkUid;
     private long _rowCount;
 
     public EpPlusSingleExcelFileWriterMutator(IEtlContext context)
